items are showing in the general menu + bug fixes

+++ lib/datas/table_orders_data/read_table_data.dart
@@ -62,7 +62,7 @@ class TableDataHandler {
     return null;
   }
 
-  Future<int> getTableTotalPrice(int tableNum) async {
+  Future<double> getTableTotalPrice(int tableNum) async {
     try {
       Map<String, dynamic>? tableSet = await getTableSet(tableNum);
